fix bugs in model && update db template

+++ src/model/entity/equipment.entity.ts
@@ -33,7 +33,7 @@ export class Equipment {
     this._name = name;
   }
 
-  @Column('int', { name: 'status' })
+  @Column('tinyint', { name: 'status' })
   public get status() {
     return this._status;
   }
